Removed Duplicated code
Moved the getTimestamp and color to a utils file and then everywhere there it was needed i just required it.
