+++ src/events/CommandEvents/evalEvent.js
@@ -1,6 +1,7 @@
 const { Events, EmbedBuilder } = require('discord.js');
 const fs = require('fs');
 const path = require('path');
+const { color, getTimestamp } = require('../../functions/utils.js');
 
 module.exports = {
     name: Events.InteractionCreate,
@@ -11,29 +12,6 @@ module.exports = {
                 const code = interaction.fields.getTextInputValue('codeInput');
                 const ephemeralInput = interaction.fields.getTextInputValue('ephemeralInput');
                 const ephemeralChoice = ephemeralInput.toLowerCase() === 'true';
-
-                const color = {
-                    red: '\x1b[31m',
-                    orange: '\x1b[38;5;202m',
-                    yellow: '\x1b[33m',
-                    green: '\x1b[32m',
-                    blue: '\x1b[34m',
-                    pink: '\x1b[38;5;213m',
-                    torquise: '\x1b[38;5;45m',
-                    purple: '\x1b[38;5;57m',
-                    reset: '\x1b[0m'
-                }
-                
-                function getTimestamp() {
-                    const date = new Date();
-                    const year = date.getFullYear();
-                    const month = date.getMonth() + 1;
-                    const day = date.getDate();
-                    const hours = date.getHours();
-                    const minutes = date.getMinutes();
-                    const seconds = date.getSeconds();
-                    return `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
-                };
 
                 if (interaction.user.id !== client.config.developers) {
                     return await interaction.reply({ content: `${client.config.ownerOnlyCommand}`, ephemeral: true,});
