+++ src/events/FixedBotStatsEvents/fixedBotStatsEvent.js
@@ -1,6 +1,7 @@
 const { Events, EmbedBuilder, version } = require('discord.js');
 const botStats = require('../../schemas/fixedBotsStatsSystem');
 const os = require('os');
+const { color, getTimestamp } = require('../../functions/utils.js');
 
 module.exports = {
     name: Events.ClientReady,
@@ -57,18 +58,3 @@ module.exports = {
         editMessage();
     }
 }
-
-const color = {
-    red: '\x1b[31m',
-}
-
-function getTimestamp() {
-    const date = new Date();
-    const year = date.getFullYear();
-    const month = date.getMonth() + 1;
-    const day = date.getDate();
-    const hours = date.getHours();
-    const minutes = date.getMinutes();
-    const seconds = date.getSeconds();
-    return `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
-};
