+++ src/events/SlashAndPrefixCreateEvents/interactionCreate.js
@@ -1,6 +1,7 @@
 const { Interaction, EmbedBuilder, ButtonBuilder, ActionRowBuilder } = require("discord.js");
 const config = require('../../config')
 const blacklistSchema = require("../../schemas/blacklistSystem");
+const { color, getTimestamp } = require('../../functions/utils.js');
 
 module.exports = {
     name: 'interactionCreate',
@@ -29,26 +30,6 @@ module.exports = {
         }
 
         const command = client.commands.get(interaction.commandName);
-
-        const color = {
-            red: '\x1b[31m',
-            orange: '\x1b[38;5;202m',
-            yellow: '\x1b[33m',
-            green: '\x1b[32m',
-            blue: '\x1b[34m',
-            reset: '\x1b[0m'
-        }
-
-        function getTimestamp() {
-            const date = new Date();
-            const year = date.getFullYear();
-            const month = date.getMonth() + 1;
-            const day = date.getDate();
-            const hours = date.getHours();
-            const minutes = date.getMinutes();
-            const seconds = date.getSeconds();
-            return `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
-        }
 
         if (!command) return
