+++ src/events/SlashAndPrefixCreateEvents/messageCreate.js
@@ -1,6 +1,7 @@
 const { EmbedBuilder, ButtonBuilder, ActionRowBuilder } = require('discord.js');
 const GuildSettings = require('../../schemas/prefixSystem');
 const blacklistSchema = require('../../schemas/blacklistSystem');
+const { color, getTimestamp } = require('../../functions/utils.js');
 
 module.exports = {
     name: "messageCreate",
@@ -46,26 +47,6 @@ module.exports = {
 
         if (!message.content.startsWith(prefix)) return;
         const args = message.content.slice(prefix.length).trim().split(/ +/);
-
-        const color = {
-            red: '\x1b[31m',
-            orange: '\x1b[38;5;202m',
-            yellow: '\x1b[33m',
-            green: '\x1b[32m',
-            blue: '\x1b[34m',
-            reset: '\x1b[0m'
-        }
-
-        function getTimestamp() {
-            const date = new Date();
-            const year = date.getFullYear();
-            const month = date.getMonth() + 1;
-            const day = date.getDate();
-            const hours = date.getHours();
-            const minutes = date.getMinutes();
-            const seconds = date.getSeconds();
-            return `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
-        }
 
         let cmd = args.shift().toLowerCase();
             if (cmd.length === 0) return;
