+++ src/functions/handlePrefix.js
@@ -1,6 +1,7 @@
 const ascii = require("ascii-table");
 const fs = require("fs");
 const table = new ascii().setHeading("File Name", "Status");
+const { color, getTimestamp } = require('../functions/utils.js');
 
 module.exports = (client) => {
     client.prefixCommands = async (eventFile, path) => {
@@ -30,26 +31,6 @@ module.exports = (client) => {
         }  
 
         client.logs.info(`[PREFIX_COMMANDS] Started refreshing prefix (?) commands.`);
-
-        const color = {
-            red: '\x1b[31m',
-            orange: '\x1b[38;5;202m',
-            yellow: '\x1b[33m',
-            green: '\x1b[32m',
-            blue: '\x1b[34m',
-            reset: '\x1b[0m'
-        }
-
-        function getTimestamp() {
-            const date = new Date();
-            const year = date.getFullYear();
-            const month = date.getMonth() + 1;
-            const day = date.getDate();
-            const hours = date.getHours();
-            const minutes = date.getMinutes();
-            const seconds = date.getSeconds();
-            return `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
-        }
 
         console.log(`${color.orange}${table.toString()} \n[${getTimestamp()}] ${color.reset}[PREFIX_COMMANDS] Loaded ${client.pcommands.size} PrefixCommands.`);
